WebApp/requirements.txt
gradio
openai-whisper
edge-tts
langchain-groq
langchain-google-genai
langchain-core
python-dotenv
